fix(deps): break circular dependency between actions/discover and benchmark/config
Move loadCliCommands and loadMcpTools from benchmark/config.ts into a new
actions/loaders.ts file, resolving the cycle:
  actions/discover → benchmark/config → project → project/snapshot → actions/discover

benchmark/config.ts re-exports both functions for backward compatibility.
