+++ src/actions/index.ts
@@ -1,4 +1,5 @@
 export { diffActionCatalog, type ActionCatalogDiff, type ChangedAction } from './diff.js';
+export { loadCliCommands, loadMcpTools } from './loaders.js';
 export { discoverActions } from './discover.js';
 export { readCliActionsFromSources } from './readers/cli.js';
 export { readMcpActionsFromSources } from './readers/mcp.js';
